Split out dbreset from dbmigrate and require passing path to migration directory.
