+++ api/cmd/dbmigrate/db-migrate-test.sh
@@ -3,11 +3,10 @@
 set -eo pipefail
 
 go build github.com/18F/e-QIP-prototype/api/cmd/dbmigrate
+go build github.com/18F/e-QIP-prototype/api/cmd/dbreset
 
 ./dbmigrate || true
 
-./dbmigrate -reset -force dbtest_test
-
-./dbmigrate dbtest_test
+./dbreset -force dbtest_test
 
 ./dbmigrate -migrations_path ./api/migrations dbtest_test
